refactor: move getExplorerUrl to src/utils
